Pangolins or scaly anteaters[1] are mammals of the order Pholidota (from the Greek word φολῐ́ς, "horny scale"). The one extant family, Manidae, has three genera: Manis, which comprises four species living in Asia; Phataginus, which comprises two species living in Africa; and Smutsia, which comprises two species also living in Africa.[2] These species range in size from 30 to 100 cm (12 to 39 in). A number of extinct pangolin species are also known.
Pangolins have large, protective keratin scales covering their skin, and they are the only known mammals with this feature. They live in hollow trees or burrows, depending on the species. Pangolins are nocturnal, and their diet consists of mainly ants and termites which they capture using their long tongues. They tend to be solitary animals, meeting only to mate and produce a litter of one to three offspring which are raised for about two years.

Pangolins are threatened by poaching (for their meat and scales) and heavy deforestation of their natural habitats, and are the most trafficked mammals in the world.[3] Of the eight species of pangolin, four (Phataginus tetradactyla, P. tricuspis, Smutsia gigantea, and S. temminckii) are listed as vulnerable, two (Manis crassicaudata and M. culionensis) are listed as endangered, and two (M. pentadactyla and M. javanica) are listed as critically endangered on the International Union for Conservation of Nature Red List of Threatened Species.[4]

Sorry I'm late Mr. White!